% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayecomp-package.R
\docType{package}
\name{bayecomp-package}
\alias{bayecomp-package}
\alias{bayecomp}
\title{The 'bayecomp' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
