% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stan.R
\name{BayeComp_Stan}
\alias{BayeComp_Stan}
\title{Bayesian Comparability with Stan}
\usage{
BayeComp_Stan(datalist, iter, chains, ...)
}
\arguments{
\item{datalist}{a list of input values.}

\item{iter}{the number of iteration.}

\item{chains}{the number of chains.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}

\item{algorithm}{algorithm for sampling.}

\item{seed}{seed value for sampling.}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian Comparability with Stan
}
